[IMP] mail: signature on reply and forward
When replying or forwarding a message from the chatter, it is still
relevant to sign the message.

We now insert the user signature below the quoted message.

Additionally fix an issue where quoted content would be cleared when
replying, even if the content wasn't part of the reply div.
Which would now impact the signature.

task-5013894
