+++ addons/mail/static/src/core/common/res_users_model.js
@@ -1,3 +1,5 @@
+import { markup } from "@odoo/owl";
+import { createElementWithContent } from "@web/core/utils/html";
 import { fields, Record } from "@mail/model/export";
 
 export class ResUsers extends Record {
@@ -19,6 +21,24 @@ export class ResUsers extends Record {
     share;
     /** @type {ReturnType<import("@odoo/owl").markup>|string|undefined} */
     signature = fields.Html(undefined);
+
+    /**
+     * Get the signature with its typical layout when inserted in html
+     */
+    getSignatureBlock() {
+        if (!this.signature) {
+            return "";
+        }
+        const divElement = document.createElement("div");
+        divElement.setAttribute("data-o-mail-quote", "1");
+        divElement.append(
+            document.createElement("br"),
+            document.createTextNode("-- "),
+            document.createElement("br"),
+            ...createElementWithContent("div", this.signature).childNodes
+        );
+        return markup(divElement.outerHTML);
+    }
 }
 
 ResUsers.register();
