+++ addons/mail/static/src/core/web/message_patch.xml
@@ -46,6 +46,12 @@
             <span>Subject: <t t-esc="message.subject || message.default_subject"/></span>
         </div>
         <t t-out="body"/>
+        <div t-if="signature" data-o-mail-quote="1">
+            <br/>
+            <t t-call="html_editor.Signature">
+                <t t-set="signatureClass" t-value="'o-signature-container'"/>
+            </t>
+        </div>
     </t>
 
     <t t-name="mail.Message.bodyInReply">
@@ -57,6 +63,12 @@
                 </div>
                 <blockquote t-out="body"/>
             </div>
+        </div>
+        <div t-if="signature" data-o-mail-quote="1">
+            <br/>
+            <t t-call="html_editor.Signature">
+                <t t-set="signatureClass" t-value="'o-signature-container'"/>
+            </t>
         </div>
     </t>
 </templates>
